#866 move create_keep_path_dir_if_missing out of hubunit class
