+++ src/a12_hub_toolbox/a12_path.py
@@ -1,8 +1,4 @@
-from src.a00_data_toolbox.file_toolbox import (
-    create_directory_path,
-    create_path,
-    set_dir,
-)
+from src.a00_data_toolbox.file_toolbox import create_directory_path, create_path
 from src.a01_term_logic.rope import get_all_rope_labels, rebuild_rope
 from src.a01_term_logic.term import BeliefLabel, BelieverName, LabelTerm
 
@@ -87,23 +83,6 @@ def create_keep_rope_path(
     keep_sub_path = create_directory_path(x_list=[*x_list])
     keeps_dir = create_keeps_dir_path(belief_mstr_dir, belief_label, believer_name)
     return create_path(keeps_dir, keep_sub_path)
-
-
-def create_keep_path_dir_if_missing(
-    belief_mstr_dir: str,
-    believer_name: BelieverName,
-    belief_label: BeliefLabel,
-    keep_rope: LabelTerm,
-    knot: str,
-):
-    keep_path = create_keep_rope_path(
-        belief_mstr_dir,
-        believer_name,
-        belief_label,
-        keep_rope,
-        knot,
-    )
-    set_dir(keep_path)
 
 
 def get_keep_dutys_path(
