+++ src/a12_hub_toolbox/test/test_hubunit/test_keep_tools.py
@@ -1,10 +1,8 @@
 from os.path import exists as os_path_exists
 from src.a01_term_logic.rope import create_rope
 from src.a05_plan_logic.plan import get_default_belief_label as root_label
-from src.a12_hub_toolbox.a12_path import (
-    create_keep_path_dir_if_missing,
-    create_keep_rope_path,
-)
+from src.a12_hub_toolbox.a12_path import create_keep_rope_path
+from src.a12_hub_toolbox.keep_tool import create_keep_path_dir_if_missing
 from src.a12_hub_toolbox.test._util.a12_env import (
     env_dir_setup_cleanup,
     get_module_temp_dir,
